auth/requirements.txt
aiohttp==3.8.3
aioredis==1.3.1
aiosignal==1.2.0
alembic==1.7.6
aniso8601==9.0.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
circuitbreaker==1.4.0
Flask==2.1.2
Flask-JWT-Extended==4.4.2
flask-marshmallow==0.14.0
marshmallow_sqlalchemy==0.28.1
Flask-Pydantic==0.9.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
httptools==0.4.0
psycopg2-binary==2.9.3
pytest==7.1.2
flasgger==0.9.5
gevent==22.10.2
redis==4.3.4
python-dotenv==0.19.2
cryptography==37.0.4
rauth==0.7.2
Flask_RESTful==0.3.9
Flask_Migrate==3.1.0
opentelemetry-api==1.14.0
opentelemetry_sdk==1.14.0
opentelemetry_instrumentation==0.35b0
opentelemetry_instrumentation_flask==0.35b0
opentelemetry_exporter_jaeger_thrift==1.14.0
